==> jupyter_notebooks_setup/3_FeatureEngineering.ipynb <==
# to handle datasets
import pandas as pd
from pandas import DataFrame
pd.pandas.set_option('display.max_columns', None)
import numpy as np

# to plot
import matplotlib.pyplot as plt
import seaborn as sns

# divide train and test set
from sklearn.model_selection import train_test_split

# feature scaling
from sklearn.preprocessing import MinMaxScaler

#### load dataset and drop the unneeded columns in the dataFrame
- rows with no gametimeplayed, datetime and the unnamed columns

ffmlDf = pd.read_csv('ffmlDF_20-21')

ffmlDf.head()

ffmlDf.shape

#### The aim of this is to predict players with the highest scores the following week, not based on OppositionTeam
- we will drop opposition team from selected_feats

# drop unneccessary columns
ffmlDf = ffmlDf.drop(columns=['Unnamed: 0','gameDate', 'Unnamed: 0.1', 'oppositionTeam'])

# remove zerominsplayed - taken from 2_DataAnalysis
def RemoveZeroMinsPlayed(df):
    df = df.copy().where(df['minsPlayed'] != 0)
    df.dropna(axis=0, inplace=True)
    return df

ffmlDf = RemoveZeroMinsPlayed(ffmlDf)

ffmlDf.head()

ffmlDf.shape

# separate data into train and test sets
X_train, X_test, y_train, y_test = train_test_split(ffmlDf,
                                                   ffmlDf['points'],
                                                   test_size=0.2,
                                                   random_state=0)
# 80:20 split
# target is df['points']
X_train.shape, X_test.shape #everything is looking good so far

Check X_train for missing or na values

var_with_na = [
    var for var in X_train.columns
    if X_train[var].isnull().sum() > 0
]
var_with_na
# if there were any vars with na we would, print the percentages
# for both the num and cat vars and fill in missing values

Check X_test for missing or na values

[var for var in X_train.columns if X_test[var].isnull().sum()]

#### Rare labels need to be identified and removed from cat_vars

- identify cat vars

cat_vars = [var for var in X_train.columns if X_train[var].dtype == 'O']
cat_vars

def FindFrequentLabels(df, var):
    df = df.copy()
    
    tmp = df.groupby(var)['points'].count() / len(df)
    
    print(tmp)
# 4480 is len


for var in cat_vars:
    FindFrequentLabels(ffmlDf, var)

As we have sooo many players the Rare percentage will have to be made a lot lower than usual.
- opp team wont be affected by this

def FindFrequentLabels(df, var, rare_perc):
    df = df.copy()
    tmp = df.groupby(var)['points'].count() / len(df)
    return tmp[tmp > rare_perc].index

# Rare Labels Romoved and Replaced!!!

for var in cat_vars:
    frequent_list = FindFrequentLabels(ffmlDf, var, 0.001)
    # i want values of 0.001 so im being a little lenient i feel
    
    X_train[var] = np.where(X_train[var].isin(frequent_list), X_train[var], 'Rare')
    X_test[var] = np.where(X_test[var].isin(frequent_list), X_test[var], 'Rare')

Hope this adds value to the model

print('X_train:',X_train['playerName'].nunique(), X_train.shape)
print('X_test:',X_test['playerName'].nunique(), X_test.shape)

I am happy with this, train and test have equall columns
- but train has more unique values (what effect will this have??)

## WE WILL NOT BE TRANSFORMING NUMERICAL VARIABLES!!!

#### Ordinal Encoding was used but causes difficulty when trying to create the Sklean pipeline
- I considered using get_dummies but due to the number of professional players it increased the features by too much
- so when it came to feature selection the, the lasso selector disregarded most of the players

# Encoding Categorical Variables

# ordinal encoding not one-hot encoding (which uses get_dummies)
def ReplaceCategories(train, test, var, target):
    # order labels in a var from low to high
    ordered_labels = train.groupby(var)[target].mean().sort_values().index
    
    # create a dictionary of ordered labels to integer values
    ordinal_label = {k: i for i, k in enumerate(ordered_labels, 0)}
    
    # use the dict to replace cat strings with integers
    train[var] = train[var].map(ordinal_label)
    test[var] = test[var].map(ordinal_label)

for var in cat_vars:
    ReplaceCategories(X_train, X_test, var, 'points')

X_train.head()

check null values in the train

[
    var for var in X_train
    if X_train[var].isnull().sum() > 0
]

check null values in test

[
    var for var in X_test
    if X_test[var].isnull().sum() > 0
]

#### Feature Scaling.

# capture all vars in a list apart from TARGET
train_vars = [
    var for var in X_train.columns
    if var not in cat_vars and var != 'points' # could have used '!='...
]
train_vars, len(train_vars)

##### Create Scaler

# Create Scaler
scaler = MinMaxScaler()
# fit scaler to the train set - never the test set
scaler.fit(X_train[train_vars])
# transform the train and test set
X_train.to_csv('xtrain.csv', index=False)
X_test.to_csv('xtest.csv', index=False)

X_train.head()
