==> jupyter_notebooks_setup/4_FeatureSelection.ipynb <==
# to handle the datasets
import pandas as pd
from pandas import DataFrame
import numpy as np
pd.pandas.set_option('display.max_columns', None)

# to plot graphs
import matplotlib.pyplot as plt
import seaborn as sns

# to build models
from sklearn.linear_model import Lasso
from sklearn.feature_selection import SelectFromModel

### Load train and test sets

X_train = pd.read_csv('xtrain.csv')
X_test = pd.read_csv('xtest.csv')

X_train.head()

#### Set the targets

y_train = X_train['points']
y_test = X_test['points']

# drop unneeded vars from the train and test sets
X_train.drop(['points'], axis=1, inplace=True)
X_test.drop(['points'], axis=1, inplace=True)

## Feature selection
We do the model fitting and feature selection altogether in a few lines of code.
- 1st we specify the Lasso Regression model,
- then we select a suitable alpha (equivalent of penalty).
- then bigger the alpha the less features will be selected.


# SelectFromModel object from sklearn is used to select features
# whose coefficients are non-zero
sel_ = SelectFromModel(Lasso(alpha=0.005, random_state=0))
# train Lasso model and select features
sel_.fit(X_train, y_train)
# visualise the features that were selected
sel_.get_support()

##### Create a list of the selected features

selected_feats = X_train.columns[(sel_.get_support())]

# print some stats
print('total features: {}'.format((X_train.shape[1])))
print('selected features: {}'.format(len(selected_feats)))
print('features with coefficients shrank to zero: {}'.format(np.sum(sel_.estimator_.coef_ ==0)))

selected_feats

compare selected_feats with the columns that were dropped

X_train.head()

#### Save the selected features

pd.Series(selected_feats).to_csv('selected_features.csv', index=False)
