==> jupyter_notebooks_setup/6_WrapUpForDeployment.ipynb <==
# to handle datasets
import pandas as pd
from pandas import DataFrame
pd.pandas.set_option('display.max_columns',None)
import numpy as np

# divide test and train
from sklearn.model_selection import train_test_split

# feature scaling
from sklearn.preprocessing import MinMaxScaler

# to build models
from sklearn.linear_model import Lasso
from sklearn.feature_selection import SelectFromModel

# to evaluate models
from sklearn.metrics import mean_squared_error, r2_score
from math import sqrt

# to persist the model and scaler
import joblib

### Load Original Data

# been a while now but load the original dataset
ffmlDf = pd.read_csv('ffmlDf_20-21')
print(ffmlDf.shape)
ffmlDf.head()

### Split into train and test set

# split into train and test
# dont forget the split used in earlier notebooks
# dont forget to set the seed
X_train, X_test, y_train, y_test = train_test_split(
    ffmlDf,
    ffmlDf['points'],
    test_size = 0.2, # 80:20 split
    random_state = 0 # setting the seed
)

X_train.shape, X_test.shape

X_train.head()

### Load the selected Features

features = pd.read_csv('selected_features.csv')
# convert to list
features = features['0'].to_list()

print('Number of feature selected:',len(features))

### Engineer Missing Values

[
    var for var in features
    if ffmlDf[var].isnull().sum() > 0
]

1. No missing values to engineer
2. No need to transform variables
3. Temporal variables dropped rather than engineered

### Rare Labels

# brought forward from notebook 3
cat_vars = [
    var for var in features
    if ffmlDf[var].dtype == 'O'
]
print(cat_vars)

def FindFrequentLabels(df, var, rare_perc):
    df = df.copy()
    tmp = df.groupby(var)['points'].count() / len(df)
    return tmp[tmp > rare_perc].index

# Rare Labels Romoved and Replaced!!!

for var in cat_vars:
    frequent_list = FindFrequentLabels(ffmlDf, var, 0.001)
    # i want values of 0.001 so im being a little lenient i feel
    print(var)
    print(frequent_list)
    print()
    
    X_train[var] = np.where(X_train[var].isin(frequent_list), X_train[var], 'Rare')
    X_test[var] = np.where(X_test[var].isin(frequent_list), X_test[var], 'Rare')

X_train[features].head()

### Encoding Categorical Variables

# code was brought forward from notebook 3_FeatureEngineering
# Encoding Categorical Variables
def ReplaceCategories(train, test, var, target):
    # order labels in a var from low to high
    ordered_labels = train.groupby(var)[target].mean().sort_values().index
    
    # create a dictionary of ordered labels to integer values
    ordinal_label = {k: i for i, k in enumerate(ordered_labels, 0)}
    
    # use the dict to replace cat strings with integers
    train[var] = train[var].map(ordinal_label)
    test[var] = test[var].map(ordinal_label)
    print(var)
    #print(ordinal_label)
    print(ordered_labels)
    print()

for var in cat_vars:
    ReplaceCategories(X_train, X_test, var, 'points')

##### Check for na in test and train

use features rather than X_train.columns or X_test.columns

[
    var for var in features
    if X_train[var].isnull().sum() > 0
]

[
    var for var in features
    if X_test[var].isnull().sum() > 0
]

### Feature Scaling

# capture the targets
y_train = X_train['points']
y_test = X_test['points']

###### Setting X_train and X_test

We need to only use X_train/test = X_train/test[features] when scaling and predicting

X_train = X_train[features] # model has to be fit to the chosen features
X_test = X_test[features] # features has to be used here as well

X_train.head()

X_test.head()

#### fit the scaler

# setup scaler
scaler = MinMaxScaler()

# fit the scaler
scaler.fit(X_train)

scaler.data_max_

scaler.data_min_

# transform the train and test set
X_train = scaler.transform(X_train)
X_test = scaler.transform(X_test)

### Train the Linear Regression: Lasso

# setup the model
# remember to set the random_state/seed
lin_model = Lasso(alpha=0.005, random_state=0)
# train the model
lin_model.fit(X_train, y_train) # X_train[features]

# persist the model for future use
joblib.dump(lin_model, 'lasso_regression.pkl')

# predict train using the model
pred = lin_model.predict(X_train)
# determine the mse, rmse and r2
print('train mse: {}'.format(int(mean_squared_error(y_train, pred))))
print('train rmse: {}'.format(int(sqrt(mean_squared_error(y_train, pred)))))
print('train r2: {}'.format(r2_score(y_train, pred)))
print()
# predict test using the model
pred = lin_model.predict(X_test)
# determine the mse, rmse and r2
print('test mse: {}'.format(int(mean_squared_error(y_test, pred))))
print('test rmse: {}'.format(int(sqrt(mean_squared_error(y_test, pred)))))
print('test r2: {}'.format(r2_score(y_test, pred)))
